requirements.txt
PyYAML~=6.0.1
spotipy~=2.23.0
pip~=24.0
redis~=5.0.2
fastapi~=0.110.0
typing_extensions~=4.10.0
sniffio~=1.3.1
requests~=2.31.0
certifi~=2024.2.2
starlette~=0.36.3
pydantic~=2.6.3
pydantic_core~=2.16.3
anyio~=4.3.0
six~=1.16.0
urllib3~=2.2.1
idna~=3.6
setuptools~=69.1.1
diskcache~=5.6.3